+++ src/communication/railroad_communication.c
@@ -1,7 +1,7 @@
 #include "communication/railroad_communication.h"
 
 
-static void send_bit_task(uint64_t message, int length)
+static void send_bit_task(unsigned long long message, int length)
 {
   outb(0x00, LPT1); // set start voltlevel
   rt_sleep(nano2count(500000)); // wait 0.5ms
@@ -34,7 +34,7 @@ if (magnetic_msg_count > 0) {
         rt_mutex_lock(&mag_sem[0]);
         MagneticTelegram telegram = buildMagneticTelegram(magnetic_msg_queue[0]);
         rt_mutex_unlock(&mag_sem[0]);
-        send_bit_task((uint64_t)telegram, length);
+        send_bit_task((unsigned long long)telegram, length);
 
         // Nachrücken
         int i;
@@ -60,7 +60,7 @@ static void send_loco_msg_task(long i)
       rt_mutex_lock(&loc_sem[i]);
       LocomotiveTelegram telegram = buildLocomotiveTelegram(locomotive_msg_queue[i]);
       rt_mutex_unlock(&loc_sem[i]);
-      send_bit_task((uint64_t)telegram, length);
+      send_bit_task((unsigned long long)telegram, length);
     }
 
     outb(0x11, LPT1);
